requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
future==1.0.0
geocoder==1.38.1
idna==3.11
numpy==2.4.2
overpy==0.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
ratelim==0.1.6
requests==2.32.5
shapely==2.1.2
simplekml==1.3.6
six==1.17.0
tkintermapview==1.29
urllib3==2.6.3
